NMSC-8 - Common code refactor.
