+++ src/app/ingredient-list/ingredient-list.component.ts
@@ -4,7 +4,7 @@ import {MatPaginator, PageEvent} from "@angular/material/paginator";
 import {BehaviorSubject, Observable, Subject, Subscription} from "rxjs";
 
 import {IngredientsService} from "../services";
-import {Ingredient} from "../models";
+import {DEFAULT_PAGE_SIZE, Ingredient, JsonBackedDataSource, resolveIcon} from "../models";
 
 @Component({
   selector: 'app-ingredient-list',
@@ -12,6 +12,8 @@ import {Ingredient} from "../models";
   styleUrls: ['./ingredient-list.component.scss']
 })
 export class IngredientListComponent implements OnInit, OnDestroy, AfterViewInit {
+
+  readonly PAGE_SIZE = DEFAULT_PAGE_SIZE;
 
   displayedColumns = ['icon', 'name', 'type', 'value'];
   dataSource: IngredientDataSource;
@@ -44,61 +46,17 @@ export class IngredientListComponent implements OnInit, OnDestroy, AfterViewInit
 
   resolveIcon(ingredient: Ingredient): string {
     const icons = this.icons.getValue();
-
-    let filename: string = null;
-    if (icons !== null) {
-      filename = icons[ingredient.name];
-    }
-
-    if (!filename) {
-      filename = ingredient.name.toLowerCase()
-        .replace(' ', '')
-        .replace("'", '');
-    }
-    return `../assets/icons/${filename}.icon.png`;
+    return `..${resolveIcon(icons, ingredient)}`;
   }
 }
 
-class IngredientDataSource extends DataSource<Ingredient> {
-
-  private data$: Subject<Ingredient[]> = new Subject<Ingredient[]>();
-  private paging: MatPaginator;
-  private subscriptions: Subscription[] = []
+class IngredientDataSource extends JsonBackedDataSource<Ingredient> {
 
   constructor(private svc: IngredientsService) {
-    super()
+    super();
   }
 
-  set paginator(paginator: MatPaginator) {
-    this.paging = paginator;
-    const sub = paginator.page
-      .subscribe( (pageEvent: PageEvent) => {
-        this.refresh(pageEvent.pageIndex, pageEvent.pageSize);
-      });
-    this.subscriptions.push(sub);
-  }
-
-  connect(): Observable<Ingredient[]> {
-    return this.data$.asObservable();
-  }
-
-  disconnect(): void {
-    this.data$.complete();
-    this.subscriptions
-      .filter(sub => !sub.closed)
-      .forEach(sub => sub.unsubscribe());
-  }
-
-  refresh(page: number = 0, pageSize: number = 5): void {
-    this.svc.getIngredients$()
-      .subscribe(all => {
-        this.paging.length = all.length;
-        this.paging.pageIndex = page;
-        this.paging.pageSize = pageSize;
-
-        const offset = page * pageSize;
-        const data = all.slice(offset, offset+pageSize);
-        this.data$.next(data);
-      });
+  protected getAllData(): Observable<Ingredient[]> {
+    return this.svc.getIngredients$();
   }
 }
