+++ src/app/recipe-list/recipe-list.component.html
@@ -29,7 +29,7 @@
   <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
 </table>
 <mat-paginator [length]="100"
-               [pageSize]="5"
+               [pageSize]="PAGE_SIZE"
                [pageSizeOptions]="[5, 10, 25, 100]">
 </mat-paginator>
